[BACKEND_MODULE_2][Development 5: Task] PR review changes: Refactor block, subject, test, task, and student test result deletion logic
- Updated deletion logic in block.helper.js to utilize CommonHelper for building delete payloads and handling subject/test deletions.
- Removed redundant functions for fetching and building delete payloads in block.helper.js, subject.helper.js, student_test_result.helper.js, and test.helper.js.
- Introduced new utility functions in shared/helper for building delete payloads and handling deletions for subjects, tests, tasks, and student test results.
- Updated validation logic to exclude 'DELETED' status for block, subject, test, and task validators.
- Changed model names to lowercase for consistency in school.model.js, student.model.js, user.model.js.
- Added detailed comments and structured the code for better readability and maintainability.
