+++ modules/block/block.helper.js
@@ -3,8 +3,9 @@ const { ApolloError } = require('apollo-server');
 
 // *************** IMPORT MODULE *************** 
 const BlockModel = require('./block.model');
-const SubjectModel = require('../subject/subject.model');
-const TestModel = require('../test/test.model');
+
+// *************** IMPORT UTILITES ***************
+const CommonHelper = require('../../shared/helper/index')
 
 // *************** IMPORT VALIDATOR ***************
 const CommonValidator = require('../../shared/validator/index');
@@ -106,11 +107,16 @@ async function GetDeleteBlockPayload({ blockId, userId }) {
         CommonValidator.ValidateObjectId(userId);
 
         const deletionTimestamp = Date.now();
-        const block = await GetBlock(blockId);
+
+        const block = await BlockModel.findOne({ _id: blockId, block_status: {$ne: 'DELETED'} });
+        if (!block) {
+            throw new ApolloError('Block not found', 'BLOCK_NOT_FOUND');
+        }
+
         const subjectIds = block.subjects || [];
 
         const deleteBlockPayload = {
-            block: BuildDeletePayload({
+            block: CommonHelper.BuildDeletePayload({
                 ids: [blockId],
                 statusKey: 'block_status',
                 timestamp: deletionTimestamp,
@@ -122,22 +128,24 @@ async function GetDeleteBlockPayload({ blockId, userId }) {
             studentTestResults: null
         };
 
+        console.log(deleteBlockPayload.block);
+
         if (!subjectIds.length) return deleteBlockPayload;
 
-        const { subjectPayload, testIds } = await HandleDeleteSubjects({ subjectIds, userId, timestamp: deletionTimestamp });
+        const { subjectPayload, testIds } = await CommonHelper.HandleDeleteSubjects({ subjectIds, userId, timestamp: deletionTimestamp });
         deleteBlockPayload.subjects = subjectPayload;
 
         if (!testIds.length) return deleteBlockPayload;
 
-        const { testPayload, taskIds, studentResultIds } = await HandleDeleteTests({ testIds, userId, timestamp: deletionTimestamp });
+        const { testPayload, taskIds, studentResultIds } = await CommonHelper.HandleDeleteTests({ testIds, userId, timestamp: deletionTimestamp });
         deleteBlockPayload.tests = testPayload;
 
         if (taskIds.length) {
-            deleteBlockPayload.tasks = HandleDeleteTasks({ taskIds, userId, timestamp: deletionTimestamp });
+            deleteBlockPayload.tasks = CommonHelper.HandleDeleteTasks({ taskIds, userId, timestamp: deletionTimestamp });
         }
 
         if (studentResultIds.length) {
-            deleteBlockPayload.studentTestResults = HandleDeleteStudentTestResults({ resultIds: studentResultIds, userId, timestamp: deletionTimestamp });
+            deleteBlockPayload.studentTestResults = CommonHelper.HandleDeleteStudentTestResults({ resultIds: studentResultIds, userId, timestamp: deletionTimestamp });
         }
 
         return deleteBlockPayload;
@@ -146,155 +154,6 @@ async function GetDeleteBlockPayload({ blockId, userId }) {
             error: error.message
         });
     }
-}
-
-/**
- * A generic utility to build a standard soft-delete deleteBlockPayload object.
- * @param {object} args - The arguments for building the payload.
- * @param {Array<string>} args.ids - An array of document IDs to target.
- * @param {string} args.statusKey - The name of the status field to be updated (e.g., 'block_status').
- * @param {number} args.timestamp - The timestamp of the deletion.
- * @param {string} args.userId - The ID of the user performing the deletion.
- * @returns {object} An object containing 'filter' and 'update' properties for a database operation.
- */
-function BuildDeletePayload({ ids, statusKey, timestamp, userId }) {
-    return {
-        filter: { _id: { $in: ids } },
-        update: {
-            [statusKey]: 'DELETED',
-            updated_at: userId,
-            deleted_by: userId,
-            deleted_at: timestamp
-        }
-    };
-}
-
-/**
- * Fetches a single block document by its ID.
- * @param {string} blockId - The ID of the block to fetch.
- * @returns {Promise<object>} A promise that resolves to the found block document.
- */
-async function GetBlock(blockId) {
-    try {
-        const block = await BlockModel.findOne({ _id: blockId, block_status: { $ne: 'DELETED' } });
-        if (!block) {
-            throw new ApolloError('Block not found', 'BLOCK_NOT_FOUND');
-        }
-        return block;
-    } catch (error) {
-        throw new ApolloError(`Failed to get block: ${error.message}`, 'GET_BLOCK_FAILED', {
-            error: error.message,
-        });
-    }
-}
-
-/**
- * Handles the processing of subject IDs for deletion, creating a payload and collecting descendant test IDs.
- * @param {object} args - The arguments for handling subject deletion.
- * @param {Array<string>} args.subjectIds - An array of subject IDs to process.
- * @param {string} args.userId - The ID of the user performing the deletion.
- * @param {number} args.timestamp - The timestamp of the deletion.
- * @returns {Promise<object>} A promise that resolves to an object containing the subject delete payload and an array of test IDs.
- */
-async function HandleDeleteSubjects({ subjectIds, userId, timestamp }) {
-    try {
-        CommonValidator.ValidateObjectIdArray(subjectIds, 'INVALID_SUBJECT_ID');
-
-        const subjects = await SubjectModel.find({ _id: { $in: subjectIds } });
-
-        if (!subjects.length) {
-            throw new ApolloError('No matching subjects found', 'SUBJECTS_NOT_FOUND');
-        }
-
-        const testIds = [].concat(...subjects.map(subject => subject.tests || []));
-
-        const subjectPayload = BuildDeletePayload({
-            ids: subjectIds,
-            statusKey: 'subject_status',
-            timestamp,
-            userId
-        });
-
-        return { subjectPayload, testIds };
-    } catch (error) {
-        throw new ApolloError(`Failed to handle delete subjects: ${error.message}`, 'HANDLE_DELETE_SUBJECTS_FAILED', {
-            error: error.message,
-        });
-    }
-}
-
-/**
- * Handles the processing of test IDs for deletion, creating a payload and collecting descendant task and result IDs.
- * @param {object} args - The arguments for handling test deletion.
- * @param {Array<string>} args.testIds - An array of test IDs to process.
- * @param {string} args.userId - The ID of the user performing the deletion.
- * @param {number} args.timestamp - The timestamp of the deletion.
- * @returns {Promise<object>} A promise that resolves to an object containing the test delete payload, an array of task IDs, and an array of student result IDs.
- */
-async function HandleDeleteTests({ testIds, userId, timestamp }) {
-    try {
-        CommonValidator.ValidateObjectIdArray(testIds, 'INVALID_TEST_ID');
-
-        const tests = await TestModel.find({ _id: { $in: testIds } });
-
-        if (!tests.length) {
-            throw new ApolloError('No matching tests found', 'TESTS_NOT_FOUND');
-        }
-
-        const taskIds = [].concat(...tests.map(test => test.tasks || []));
-        const studentResultIds = [].concat(...tests.map(test => test.student_test_results || []));
-
-        const testPayload = BuildDeletePayload({
-            ids: testIds,
-            statusKey: 'test_status',
-            timestamp,
-            userId
-        });
-
-        return { testPayload, taskIds, studentResultIds };
-    } catch (error) {
-        throw new ApolloError(`Failed to handle delete tests: ${error.message}`, 'HANDLE_DELETE_TESTS_FAILED', {
-            error: error.message,
-        });
-    }
-}
-
-/**
- * Handles the processing of task IDs for deletion and creates the corresponding payload.
- * @param {object} args - The arguments for handling task deletion.
- * @param {Array<string>} args.taskIds - An array of task IDs to process.
- * @param {string} args.userId - The ID of the user performing the deletion.
- * @param {number} args.timestamp - The timestamp of the deletion.
- * @returns {object} An object containing the 'filter' and 'update' payload for tasks.
- */
-function HandleDeleteTasks({ taskIds, userId, timestamp }) {
-    CommonValidator.ValidateObjectIdArray(taskIds, 'INVALID_TASK_ID');
-
-    return BuildDeletePayload({
-        ids: taskIds,
-        statusKey: 'task_status',
-        timestamp,
-        userId
-    });
-}
-
-/**
- * Handles the processing of student test result IDs for deletion and creates the corresponding payload.
- * @param {object} args - The arguments for handling student test result deletion.
- * @param {Array<string>} args.resultIds - An array of student test result IDs to process.
- * @param {string} args.userId - The ID of the user performing the deletion.
- * @param {number} args.timestamp - The timestamp of the deletion.
- * @returns {object} An object containing the 'filter' and 'update' payload for student test results.
- */
-function HandleDeleteStudentTestResults({ resultIds, userId, timestamp }) {
-    CommonValidator.ValidateObjectIdArray(resultIds, 'INVALID_STUDENT_TEST_RESULT_ID');
-
-    return BuildDeletePayload({
-        ids: resultIds,
-        statusKey: 'student_test_result_status',
-        timestamp,
-        userId
-    });
 }
 
 // *************** EXPORT MODULE ***************
