+++ modules/studentTestResult/student_test_result.helper.js
@@ -4,6 +4,9 @@ const { ApolloError } = require('apollo-server');
 // *************** IMPORT MODULE *************** 
 const StudentTestResultModel = require('./student_test_result.model');
 
+// *************** IMPORT UTILITES ***************
+const CommonHelper = require('../../shared/helper/index')
+
 // *************** IMPORT VALIDATOR ***************
 const CommonValidator = require('../../shared/validator/index');
 const StudentTestResultValidator = require('./student_test_result.validator');
@@ -47,11 +50,16 @@ async function GetDeleteStudentTestResultPayload({ studentTestResultId, userId }
         CommonValidator.ValidateObjectId(userId);
 
         const deletionTimestamp = Date.now();
-        const studentTestResult = await GetStudentTestResult(studentTestResultId);
+
+        const studentTestResult = await StudentTestResultModel.findOne({ _id: studentTestResultId, student_test_result_status: { $ne: 'DELETED' } });
+        if (!studentTestResult) {
+            throw new ApolloError('Student test result not found', 'STUDENT_TEST_RESULT_NOT_FOUND');
+        }
+
         const testId = studentTestResult.test;
 
         const deletePayload = {
-            studentTestResult: BuildDeletePayload({
+            studentTestResult: CommonHelper.BuildDeletePayload({
                 ids: [studentTestResultId],
                 statusKey: 'student_test_result_status',
                 timestamp: deletionTimestamp,
@@ -66,47 +74,6 @@ async function GetDeleteStudentTestResultPayload({ studentTestResultId, userId }
             error: error.message
         });
     }
-}
-
-/**
- * Fetches a single student test result document by its ID.
- * @param {string} studentTestResultId - The ID of the student test result to fetch.
- * @returns {Promise<object>} A promise that resolves to the found student test result document.
- */
-async function GetStudentTestResult(studentTestResultId) {
-    try {
-        const studentTestResult = await StudentTestResultModel.findOne({ _id: studentTestResultId, student_test_result_status: { $ne: 'DELETED' } });
-        if (!studentTestResult) {
-            throw new ApolloError('Student test result not found', 'STUDENT_TEST_RESULT_NOT_FOUND');
-        }
-
-        return studentTestResult;
-    } catch (error) {
-        throw new ApolloError(`Failed to get student test result: ${error.message}`, 'GET_STUDENT_TEST_RESULT_FAILED', {
-            error: error.message
-        });
-    }
-}
-
-/**
- * A generic utility to build a standard soft-delete payload object.
- * @param {object} args - The arguments for building the payload.
- * @param {Array<string>} args.ids - An array of document IDs to target.
- * @param {string} args.statusKey - The name of the status field to be updated.
- * @param {number} args.timestamp - The timestamp of the deletion.
- * @param {string} args.userId - The ID of the user performing the deletion.
- * @returns {object} An object containing 'filter' and 'update' properties for a database operation.
- */
-function BuildDeletePayload({ ids, statusKey, timestamp, userId }) {
-    return {
-        filter: { _id: { $in: ids } },
-        update: {
-            [statusKey]: 'DELETED',
-            updated_by: userId,
-            deleted_by: userId,
-            deleted_at: timestamp
-        }
-    };
 }
 
 /**
