+++ modules/user/user.model.js
@@ -80,7 +80,7 @@ const userSchema = mongoose.Schema({
     }
 });
 
-const UserModel = mongoose.model('User', userSchema);
+const UserModel = mongoose.model('user', userSchema);
 
 // *************** EXPORT MODULE ***************
 module.exports = UserModel;
